requirements.txt
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
six==1.16.0
SQLAlchemy==1.4.21
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
